fullstack: Fix race between updating port security and test
Patch I65da39fd390e4faecc6cfb18bb50e1f5ce684f1e changed the test to
start with port-security-enabled set to False and with no security
group. Later port security is turned on and security group is given but
immediatelly starts testing connectivity. It may happen that security
groups are not applied to the port yet and negative connectivity test
(former No. 3) fails. This patch actively waits for connection not to
pass and then continues as before.

Change-Id: Iddd66770638973a0b3c916f068b5c813886e07ef