Move dhcp_release6_supported to runtime checks file
Since one of the main reasons for the sanity check system was to
avoid doing runtime checks, importing a sanity check to do a
runtime check encourages bad behavior. This moves the check to a
new runtime_checks.py file that includes a note encouraging people
to use sanity checks wherever possible.

Change-Id: I06bffe00bb796b4727dca7867a15302582ffcc10