Registry decorator to bring subscribes to methods
This adds some sugar to allow subscription to resources via
a decorator directly on a method as it is defined in a class.

This will allow us to stop overriding __new__ everywhere for
subscriptions and will allow the fact that a method is subscribed
to a resource to be visible right above the method itself.

See the follow-up patch[1] for an example of how this can clean
things up.
1. I764be99b2edbd0e7b54673c1940e883d7f313bf5

Change-Id: I1020c5ddec57f535e5b94f803144dfa19d668937