functional: Stop compiling OVS from source
There is a bug 1640283 fixed in OVS 2.5.1 but currently we ship Neutron
with openvswitch 2.5.0. The patch adds a decorator that skips particular
tests in case minimal version requirement is not met.

Mitaka changes: test case not skipped because there is no such test case
in the branch.

Change-Id: I3a665f3ba770e4acad7e7ead3f8cc557a86952cf
(cherry picked from commit 7210696702493736ea3d203fa13bb13012207632)