Convert gate_hook to devstack-tools
Devstack does not support multiple duplicate sections in local.conf.
Currently it works by chance, and is not guaranteed to work when next
changes land for devstack. So we should make sure we generate local.conf
that does not have duplicate sections.

This can be achieved using tools from devstack-tools package, that allow
to produce and merge multiple local.conf files compatible with devstack.

This change will need to be backport to all stable branches to
accommodate for branchless devstack-gate repo.

Related-Bug: #1667331
Co-Authored-By: Ihar Hrachyshka <ihrachys@redhat.com>
Depends-On: I53119876b82eeaec19382dbe115cb1740bed0cec
Change-Id: Ibd0f67f9131e7f67f3a4a62cb6ad28bf80e11bbf