Fix has_registry_receivers for classes w/o __new__
When a class doesn't define __new__, the __new__ method is inherited
from the parent class using MRO; however, we capture the orig_new at
class definition time which doesn't have the definitive MRO because
we don't know how the class will be used. So calling orig_new when
it's inherited ends up skipping sibling class __new__ methods.

This fixes the has_registry_receivers decorator to detect the case
when a class doesn't have a custom __new__ method and performs
a super() call like a custom __new__ method would to ensure we
don't miss calling __new__ on any classes in the heirarchy.

Change-Id: I547a887dac113625a1045d95bab099059e99fc0c