Bump to Nova V2.1
Nova V2.1 API has been available since Kilo, and the API is CURRENT.
Neutron is using Nova API via novaclient and the used Nova API is
V2.1 on the gate as the default Nova endpoint. So this patch doesn't
change the used Nova API actually, but current novaclient skips checking
expected API version in a response from Nova API.
This patch will be helpful to avoid an issue when enabling this check
on novaclient side.

NOTE: Cinder is using Nova v2.1 API with the similar patch as
      Iadd3363265be6c5a8ed46704a712da6c15e2b046

Change-Id: Ie3a10f7078f4213da77be522e3866f82b5e8683c