Don't IPAddressGenerationFailure for DHCP ports
This is a runtime state users can easily create by creating tiny subnets
with no room for DHCP agent addresses. It's not something we need to
dump tracebacks on the server for because it's not a bug or unexpected
outcome.

These traces can be seen all over the API test logs where small subnets
are created without the need to actuall put things on them.

TrivialFix

Change-Id: I663b0df07608d8511a40c82e589875c50c7a6066