Stabilizing process monitor function test case
The problem seems to be that the TC is starting new processes
parallelly but it does not wait for them to start properly instead
it tries to kill one of them immediately but in some cases if the
startup phase is delayed the process that is about to be killed is
not started yet.
The fix is to introduce a waiting phase until all the child
processes are up and running right after the initial spawning
phase.
As the problem is not functional the fix is difficult to test.
I executed this TC in a loop with this fix for a 5000 times and
the problem has not appeared. Without the fix I had 3.8% failure
rate.

Change-Id: I19a8a4c35efc35a002478204fd24910d73c9c8c2
closes-bug: #1513771