Make RBAC entry removal enginefacade friendly
The previous approach of deleting the record and then expiring
the rbac_entries relationship was triggering a refresh from the
DB during make_network_dict. This pattern will break once we switch
to the new enginefacade so this patch just adjusts the logic to
remove the entry via the relationship.

Partially-Implements blueprint: enginefacade-switch

Change-Id: I74c88f98c1f5af4d505d17967d1c15745fe225ec