Fullstack: enable DHCP agent
With the merge of I8111469ad4b0d88580bff7a77492ad95af8e9377,
ports will not transition to the ACTIVE status on a subnet
that has DHCP enabled that does not clear the DHCP provisioning
block. So if we don't deploy a DHCP agent for these tests,
the ports will not transition to ACTIVE.

This fixes the fullstack tests by adding DHCP agents
to clear the provisioning block.

Co-Authored-By: Jakub Libosvar <libosvar@redhat.com>
Closes-bug: #1677729
Change-Id: I66c5e4ead23f0e6bd4b9e867c8cf419f67f6d78c