Add work-around for older oslo.db
The min version of oslo.db in Newton at the time of this
commit is 4.10 and Ocata is 4.15, which do not include
a commit [1] to skip None values in pagination.
Rather than trying to get oslo.db globally bumped to
4.17.0 on both stable branches, this just skips None
in our code.

This is not proposed to master because it uses a new enough
oslo.db to not be impacted by this.

1. b3869d04cff7071c1226758eb8b58fde9eba5b8d

Closes-Bug: #1680136
Change-Id: Ieef10eebd93f99404dd2fd87ccbab9b75632945a