Don't check for enable_security_group vs. firewall_driver compatibility
This is something really old, like... Havana old. This code is also
triggered by neutron-server when loading extension for SGs, which
doesn't make sense at all, because server is not supposed to configure
firewall_driver (it's agent job). So in setups that (correctly) leave
firewall_driver unset in neutron.conf, we get a warning log message.

It's not the only guilt of the check: it also compares firewall_driver
against full import path to noop driver instead of using stevedore
aliases.

This patch just kills the whole check, from both agent as well as server
sides.

Change-Id: Iec9893cb2e30967a77f4f7151d5c31d45cd8e690