Call expire_all in revision tests before re-use
After the merge of Ia15c63f94d2c67791da3b65546e59f6929c8c685, delete
port occurs in a separate session. So when an interface is deleted
from a router, the router object that gets a revision bump is in a
separate sqlalchemy session.

This means that if the router object is loaded in a parent session,
it will not see the updated revision number. This is fine for the normal
code path since we don't return the revision number of the router after
the port removal in the API. However, in the unit tests we keep re-using
the same session for all of the operations so we can get a stale router
in a get_router call after an interface removal if the router SQLA object
is still in the identity map of the session.

The transience of this failure was due to the fact that a stale read
would only occur if the router SQLA object hadn't been garbage collected.

This fix alters the unit tests to call expire_all() on the session being
used whenver it is referenced if it's not in an active transaction.

Change-Id: I5ba5c5773b85ed9a710e69d7e2f596f760b43fd3
Closes-Bug: #1679815