segments: make sure we pass lists into get_objects
get_objects doesn't accept dict_keys objects that sometimes may be
passed in by consumers of the code (for example, ovn does just that).
To make sure we still work for those consumers without a change on their
side, enforce conversion to a list before passing the list of physical
network names down the stack.

Depends-On: Ied0516686e86167a0c7b1d480eb1db2789f7cada
Change-Id: Ic6c5220cf3f1e901e159e0b76534155985ce9fe2