Fixup event transaction semantics for ML2 bulk ops
This splits out code in ML2 for ports, subnets, and networks
into clear before, during, and after functions corresponding
to the state of the transaction they are called within.
This allows the bulk creation to correctly call chunks of code
outside of DB transactions even when doing many creations at once.

Change-Id: Id6fe0f5066358be954a6ca14dd49c36755897e31
Closes-Bug: #1683550