Split out DVR DB into multiple classes
This splits the DVR class into subclasses. One that contains the logic
that will ultimately be independent from the L3 mixin which contains the
callbacks that will be used by the DVR service provider. The second
contains all of the agent and scheduling interface logic that will
likely need to remain a mixin for the foreseeable future.

Related: blueprint multi-l3-backends
Change-Id: I64051c9850c82407801cfd9bf9b4ddba3374b9ac