Clean MissingAuthPlugin from unit tests
In the unit test logs there is a MissingAuthPlugin exception
printed several times.

This patch aims to mock the notifier when it's not being tested.
When it is being tested, like in the NovaSegmentNotifierTestCase,
mocking the keystone auth prevents the exception from being
thrown.

Closes-Bug: #1683606
Change-Id: I63a1c887f6a7a43f2bddd48b86078e5836cbf438
(cherry picked from commit ac527e242938af280fc26b11b572767177837add)