Fix security group rules created for dhcpv6
In [1] the setup of security group rules for DHCP was refactored.
However it assumed that DHCPv6 responses would go to the target IP,
instead the destination will be the link-local address of the instance,
so we will need to add this to the list of allowed destinations.

[1] Ibfbf011284cbde396f74db9d982993f994082731

Change-Id: I6bd24a98d7c72a1bac28999eb3c9a831ab2ab9bf
Closes-Bug: 1693771
(cherry picked from commit 1e9aca11b9736993e39972a36ecd53765f4b580a)