Do not respond to ARP on IPv6-only interfaces
When using dual stack, the IPv6 router interface responds
to ARP requests that only the IPv4 interface should.
This results in ARP flux and can cause a guest to address
packets to the wrong layer-2 address when sending traffic
to the IPv4 gateway.

Change arp_ignore and arp_announce sysctl options on interfaces
in the router namespace to be more strict in how we respond.

Closes-bug: 1692007
Change-Id: Ic3c2370995abb027a3412b473ce6bc63790c1105