objects: exclude revision_number from updatable fields
Revision service plugin is the entity that updates the database model
attribute, and we don't allow to modify the attribute on API layer. We
should not allow to explicitly update it on OVO layer either.

This can help with spurious unit test failures that sometimes try to
update revision number and then reuse the same number to fetch the
object. If revision service plugin decides to bump the number during the
test case execution once more, using the old number may yield no
results.

Change-Id: I4666c672223ac38508738b739207f754847c62bc
Closes-Bug: #1694753