Stop arping when interface gets deleted
It is possible for an interface to be added to a
router, have arping get started for it in a thread,
then immediately remove the interface, causing
arping errors in the l3-agent log.  This concurrent
deletion should be handled more gracefully by
just logging a warning on the first detection and
returning early.

Ocata changes: added translation marker to a warning message because
it's enforced in the branch. (In Pike, we reversed the direction, not
adding those markers to log messages.)

Change-Id: I615b60561b3b7f8c950d5f412fb4cdf7877b98f7
Closes-bug: #1696893
(cherry picked from commit 739daaa9555e734b94bef89f4fe1c5159c8fd435)