DVR: Fix neutron metering agent to notify hosts hosting DVR
DVR router updates are notified based on where they are hosted.
But in the case of metering rpc notifier the notification to a
specific host was not implemented and also when a call is being
made to routers_updated_on_host, it is being ignored since it only
implements the routers_updated function call.

Change-Id: I39b0d43b14294a8eecf6cba230d948c2c45a0b7a
Closes-Bug: #1682345
(cherry picked from commit 72338d9fad620ea5615a283d1e1d3cb7b56a5d9d)