Always try to delete bridge for ID on network_delete
If network_deletes are received before port creates
are processed, the agent might not have the network in
it's map even though it has a bridge to delete.

This adjusts the logic to always try to delete the bridge
corresponding to a network_id even if it's not in the
network_map yet.

Change-Id: I5e72bff2ffd9568f272ed48187ad543ab5a3d1ec
Closes-Bug: #1698271