Add revises_on_change to Binding DB models
All of the ML2 binding models were missing the revises_on_change
attribute to bump the revision of the port whenever they are
created/updated/deleted.

These are important because port binding happens in a separate
transaction so without them the revision number before and after
port binding was exactly the same. This opened up the L2 agent
to a race condition dependent on the order of processed push
notifications.

Closes-Bug: #1699034
Change-Id: I75f3c63941130ce845574e60214ac34e99111693