Replaced assertTrue(False) with fail()
In some unit and functional tests self.assertTrue(False) was used
instead of self.fail(), which might be against readability.

Using assertTrue(False) gives the following message on fail:

  File "C:\Python361\lib\unittest\case.py", line 678, in assertTrue
    raise self.failureException(msg)
    AssertionError: False is not true

After replacing assertTrue(False) with fail():

  File "C:\Python361\lib\unittest\case.py", line 666, in fail
    raise self.failureException(msg)
    AssertionError: None

Although the results are the same (both tests failed), the
message 'False is not true' is unnecessary, and can be omitted
from the log by using fail().

Change-Id: I81e21040fd6a2f9713889912fafd2b19bd056b5a